pyproject.toml
[project]
name = "fine_tuning"
version = "0.1.0"
description = "Fine tune a model to scan bills"
readme = "README.md"
authors = [
    {name = "José Miguel Vilches", email = "josemiguelvilchesfierro@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "hf-transfer>=0.1.9",
    "loguru>=0.7.3",
    "modal>=1.1.4",
    "peft>=0.17.1",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
    "trl>=0.23.0",
    "wandb>=0.21.4",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
include = ["services", "."]
extraPaths = ["."]          # This helps with imports
